Move HandleSubscription et al. around per Arko's feedback
I was going to do a type alias for `watchable.Update`, but:

    internal/message/watchutil.go:7:6: generic type cannot be alias

So I just defined a new child type, which is fine because there aren't any
methods on Update.
