+++ internal/provider/kubernetes/gateway.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 
 	"github.com/go-logr/logr"
-	"github.com/telepresenceio/watchable"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
@@ -30,7 +29,6 @@ import (
 	"github.com/envoyproxy/gateway/internal/provider/utils"
 	"github.com/envoyproxy/gateway/internal/status"
 	"github.com/envoyproxy/gateway/internal/utils/slice"
-	"github.com/envoyproxy/gateway/internal/utils/watchutil"
 )
 
 const gatewayClassFinalizer = gwapiv1b1.GatewayClassFinalizerGatewaysExist
@@ -624,8 +622,8 @@ func (r *gatewayReconciler) envoyDeploymentForGateway(ctx context.Context, gatew
 // Kubernetes API Server
 func (r *gatewayReconciler) subscribeAndUpdateStatus(ctx context.Context) {
 	// Subscribe to resources
-	watchutil.HandleSubscription(r.resources.GatewayStatuses.Subscribe(ctx),
-		func(update watchable.Update[types.NamespacedName, *gwapiv1b1.Gateway]) {
+	message.HandleSubscription(r.resources.GatewayStatuses.Subscribe(ctx),
+		func(update message.Update[types.NamespacedName, *gwapiv1b1.Gateway]) {
 			// skip delete updates.
 			if update.Delete {
 				return
